Move adapters para módulo de webservices
